add watch flag to test command
