+++ cmd/test.go
@@ -11,6 +11,7 @@ import (
 
 func testCommand() (*cobra.Command, error) {
 	v := viper.New()
+	var watch bool
 
 	root := &cobra.Command{
 		Use:   "test [PATH]",
@@ -38,7 +39,10 @@ Complete documentation at https://mlange-42.github.io/modo/`,
 			if err != nil {
 				return err
 			}
-			return runTest(cliArgs)
+			if err := runTest(cliArgs); err != nil {
+				return err
+			}
+			return watchAndRun(cliArgs, runTest)
 		},
 	}
 
@@ -48,6 +52,7 @@ Complete documentation at https://mlange-42.github.io/modo/`,
 	root.Flags().BoolP("strict", "S", false, "Strict mode. Errors instead of warnings")
 	root.Flags().BoolP("dry-run", "D", false, "Dry-run without any file output")
 	root.Flags().BoolP("bare", "B", false, "Don't run ore- and post-commands")
+	root.Flags().BoolVarP(&watch, "watch", "W", false, "Watch for changes to sources and documentation files")
 	root.Flags().StringSliceP("templates", "T", []string{}, "Optional directories with templates for (partial) overwrite.\nSee folder assets/templates in the repository")
 
 	root.Flags().SortFlags = false
